Changed recombination_exp()
Change list (for later discussions with Moises):

- Changed “@section Recombination Parameters:” (roxygen documentation):
Always verify that the correct fields in ‘recpars’ are documented. The
“/exp” recombination requires just one parameter, ‘cr’ (element-wise
probability of selection as cut-point)

- Changed function definition (line 23):
Input list “recpars” does not contain a parameter “K”, but instead
should contain a parameter “cr”. Since there is no particular
justification for giving “cr” a standard value, we do not define a
value in the function definition

- Included error catching for the existence of “recpars$cr” (line 28)

- Removed function makeinherits() (it was wrong, as I mentioned
yesterday). Instead, I defined two vectors, one uniformly distributed
for the start point of the recombination (mut.start) and another
exponentially distributed for the end point (mut.end), with one element
for each individual (see lines 34 - 45)

- Added helper function setfun(n,s,e) (line 51)
This function returns a vector with value 1 in all positions between a
start point (s) and an end point (e mod n), wrapping around the end of
a vector of length n. All other positions are set to 0.

- Used mapply() to set a recombination matrix similar to the one used
in recombination_bin() (line 58)
mapply() iterates over multiple vectors, so we can use it to generate a
recombination matrix that uses different start and end points for each
individual.

As you can see, I completely changed the function. Please read and
*understand* (it’s well commented) what I did, so that you can build
similar ones in the future.